% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drtm.R
\name{calculate_energy}
\alias{calculate_energy}
\title{Global energy:}
\usage{
calculate_energy(idx, seg, pop, walk, bicy)
}
\arguments{
\item{idx}{numeric, indices of candidates ins 'seg'.}

\item{seg}{sf, road segments point locations.}

\item{pop}{sf, population point data.}

\item{walk}{dodgr routing graph, graph to route walking times between the points.}

\item{bicy}{dodgr routing graph, graph to route bicycle times between the points.}
}
\value{
A data.table object containing the energies.
}
\description{
Global energy:
}
\examples{
print("tbd.")
}
