% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drtm.R
\name{drt_drtm}
\alias{drt_drtm}
\title{Model class definition}
\usage{
drt_drtm(
  model_name,
  aoi,
  pop,
  n_vir,
  m_seg = 100,
  calc_energy = e_walkDrive_pop
)
}
\arguments{
\item{model_name}{character, name of the drtm.}

\item{aoi}{sf object, polygon of the Area of Interest (AOI).}

\item{pop}{sf object, centroids of a hectaraster population dataset covering the full extent of the \code{aoi} input (column name for population must be 'n').}

\item{n_vir}{numeric, number of the virtual stations to place.}

\item{m_seg}{numeric, resolution of the road segmentation in meters.}

\item{calc_energy}{function, energy calculation function.}
}
\value{
A demand responsive transport model of class \code{drtm}.
}
\description{
Demand responsive transport model.
}
\examples{
m <- drt_drtm(
  model_name = "Jegenstorf",
  aoi = aoi, pop = pop,
  n_vir = 10, m_seg = 100
)
}
