% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drtm.R
\name{drt_energy<-}
\alias{drt_energy<-}
\title{Set energy function}
\usage{
drt_energy(x) <- value
}
\arguments{
\item{x, }{drtm, a drtm model.}

\item{value}{function, energy function to calculate energy of the model.}
}
\value{
A new drtm with the new energy.
}
\description{
Set energy function
}
\examples{
# Example model
m <- drt_import(
  system.file("Jegenstorf_i1000.RData", package = "drtplanr")
)

drt_energy(m) <- calculate_energy
}
