% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drtm.R
\name{drt_energy}
\alias{drt_energy}
\title{Get energy function}
\usage{
drt_energy(x)
}
\arguments{
\item{x, }{drtm, a drtm model.}
}
\value{
The energy function of the model.
}
\description{
Get energy function
}
\examples{
# Example model
m <- drt_import(
  system.file("Jegenstorf_i1000.RData", package = "drtplanr")
)

drt_energy(m)
}
