% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drtm.R
\name{drt_export}
\alias{drt_export}
\alias{drt_export.drtm}
\title{Export model}
\usage{
drt_export(x, ...)

\method{drt_export}{drtm}(obj, path)
}
\arguments{
\item{obj}{drtm, a drtm model.}

\item{path}{character, path to file.}
}
\value{

}
\description{
Export model
}
