% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drtm.R
\name{drt_import}
\alias{drt_import}
\alias{drt_import.character}
\title{Import model}
\usage{
drt_import(x, ...)

\method{drt_import}{character}(file_name)
}
\arguments{
\item{file_name}{character, path to file.}
}
\value{

}
\description{
Import model
}
