% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drtm.R
\name{drt_iterate}
\alias{drt_iterate}
\alias{drt_iterate.drtm}
\title{Iterate}
\usage{
drt_iterate(obj, ...)

\method{drt_iterate}{drtm}(obj, n_iter)
}
\arguments{
\item{obj, }{drtm, a drtm model.}

\item{n_iter}{numeric, number of iterations.}
}
\value{
A new \code{drtm} object with \code{n_iter} times more iterations.
}
\description{
Minimize the gloabal energy of a drtm model.
}
\examples{
m <- drt_drtm(
model_name = "Jegenstorf",
aoi = aoi, pop = pop,
n_vir = 10, m_seg = 100
)

drt_iterate(m, 10)
}
