% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drtm.R
\name{drt_map}
\alias{drt_map}
\alias{drt_map.drtm}
\title{Plot station map}
\usage{
drt_map(x, ...)

\method{drt_map}{drtm}(obj)
}
\arguments{
\item{obj}{drtm, a drtm model.}
}
\value{

}
\description{
Plot station map
}
