% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drtm.R
\name{drt_mask}
\alias{drt_mask}
\title{Mask a spatial layer by a polygon}
\usage{
drt_mask(layer, aoi)
}
\arguments{
\item{layer}{spatial object to crop by the polygon.}

\item{aoi}{spatial object, polygon to mask to.}
}
\value{
A masked sf.
}
\description{
Mask a spatial layer by a polygon
}
\examples{
# Example data
aoi <-
  sf::st_read(system.file("example.gpkg", package = "drtplanr"), layer = "aoi")

pop <-
  sf::st_read(system.file("example.gpkg", package = "drtplanr"), layer = "pop")

drt_mask(pop, aoi)
}
