% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drtm.R
\name{drt_reset}
\alias{drt_reset}
\title{Reset the model state}
\usage{
drt_reset(obj, n_sta, shuffle = FALSE)
}
\arguments{
\item{obj, }{drtm, a drtm model.}

\item{n_sta, }{numeric, number of stations (default = NULL).}

\item{shuffle, }{boolean, shuffle the initial station positions?}
}
\value{
The energy function of the model.
}
\description{
Reset the model state
}
\examples{
# Example model
m <- drt_import(
  system.file("Jegenstorf_i1000.RData", package = "drtplanr")
)

drt_reset(m)
}
