% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drtm.R
\name{drt_save_graphics}
\alias{drt_save_graphics}
\title{Save graphics of model}
\usage{
drt_save_graphics(obj, path)
}
\arguments{
\item{obj}{drtm, a drtm model.}

\item{path}{character, path to file.}
}
\value{
None.
}
\description{
Save graphics of model
}
\examples{
#' # Example model
m <- drt_import(
  system.file("Jegenstorf_i1000.RData", package="drtplanr")
)

# Save to temp dir
drt_save_graphics(m, path = tempdir())
}
