% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drtm.R
\name{drt_summary}
\alias{drt_summary}
\title{Reset the model state}
\usage{
drt_summary(obj)
}
\arguments{
\item{obj, }{drtm, a drtm model.}

\item{shuffle, }{boolean, shuffle the initial station positions?}
}
\value{
The energy function of the model.
}
\description{
Reset the model state
}
\examples{
# Example model
m <- drt_import(
  system.file("Jegenstorf_i1000.RData", package = "drtplanr")
)

drt_energy(m)
}
