% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drtplanr.R
\docType{package}
\name{drtplanr-package}
\alias{drtplanr-package}
\title{Demand-responsive transport planner}
\description{
Lorem ipsum
}
\author{
Merlin Unterfinger - \email{info@munterfinger.ch}
}
\keyword{package}
