% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drtm.R
\name{e_walk_bike}
\alias{e_walk_bike}
\title{Global energy: sum(walk*bike/pop)}
\usage{
e_walk_bike(idx, seg, pop, graph)
}
\arguments{
\item{idx}{numeric, indices of candidates ins 'seg'.}

\item{seg}{sf, road segments point locations.}

\item{pop}{sf, population point data.}

\item{graph}{dodgr routing graph, graph to route between the points.}
}
\value{
A data.table object containing the energies.
}
\description{
Global energy: sum(walk*bike/pop)
}
\examples{
print("tbd.")
}
