% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drtm.R
\name{e_walkDrive_pop}
\alias{e_walkDrive_pop}
\title{Global energy: sum((walk+5*drive)/pop)}
\usage{
e_walkDrive_pop(idx, seg, pop, graph)
}
\arguments{
\item{graph}{}
}
\value{

}
\description{
Global energy: sum((walk+5*drive)/pop)
}
