% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drtm.R
\name{print.drtm}
\alias{print.drtm}
\title{Print}
\usage{
\method{print}{drtm}(x, ...)
}
\arguments{
\item{x}{drtplanr object, print information about package classes.}

\item{...}{...}
}
\value{
None.
}
\description{
Print
}
\examples{
# Example model
m <- drt_import(
  system.file("Jegenstorf_i0.RData", package="drtplanr")
)
print(m)
}
