% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drtplanr-package.R
\name{tmessage}
\alias{tmessage}
\title{Message with timestamp}
\usage{
tmessage(text)
}
\arguments{
\item{text}{character, test of the message}
}
\value{
None.
}
\description{
Message with timestamp
}
\examples{
tmessage("Test")
}
