% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drtm.R
\name{calculate_energy}
\alias{calculate_energy}
\title{Global energy of the model}
\usage{
calculate_energy(idx, seg, pop, walk, bicy)
}
\arguments{
\item{idx}{numeric, indices of candidates ins 'seg'.}

\item{seg}{sf, road segments point locations.}

\item{pop}{sf, population point data.}

\item{walk}{dodgr routing graph, graph to route walking times between the points.}

\item{bicy}{dodgr routing graph, graph to route bicycle times between the points.}
}
\value{
A numeric scalar with the global energy.
}
\description{
Energy function that calculates the global energy of the model by summing
product of the attractivities of each origin and destination station pair
weighted by their connectivity (bicycle time between the stations).
}
\details{
Notation:
\itemize{
 \item{Ps_i: }{Sum of population in catchment area of station i.}
 \item{WTs_i: }{Sum of the walking time of population in catchment area to station i.}
 \item{BTs_i_s_j: }{Bicycle time from station i to station j.}
 \item{LEs_i = Ps_i * WTs_i: }{Local energy of station i.}
}

Terms:
\itemize{
 \item{s1 = }{sum(Ps_i * LEs_i)}
 \item{s2 = }{sum(Ps_i * BTs_i_s_j * Ps_j)}
 \item{s3 = }{sum(Ps_j * LEs_j)}
}

Global energy: e_g = sum(s1 + s2 + s3)
}
\examples{
print("tbd.")
}
