% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drtm.R
\name{drt_export}
\alias{drt_export}
\title{Export drtm}
\usage{
drt_export(obj, path)
}
\arguments{
\item{obj}{drtm, a drtm model of the drtplanr.}

\item{path}{character, path to file.}
}
\value{
None.
}
\description{
Export drtm
}
\examples{
\donttest{
# Example model
m <- drt_import(
  system.file("Buelach_i1000.RData", package = "drtplanr")
)

# Export to temporary dir
drt_export(m, path = tempdir())
}
}
