% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drtm.R
\name{drt_import}
\alias{drt_import}
\title{Import drtm}
\usage{
drt_import(file_name)
}
\arguments{
\item{file_name}{character, path to file.}
}
\value{
A drtm model.
}
\description{
Import drtm
}
\examples{
# Example model
m <- drt_import(
  system.file("Jegenstorf_i0.RData", package="drtplanr")
)
}
