% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drtm.R
\name{drt_iterate}
\alias{drt_iterate}
\title{Iterate}
\usage{
drt_iterate(obj, n_iter, annealing = TRUE)
}
\arguments{
\item{obj, }{drtm, a drtm model.}

\item{n_iter}{numeric, number of iterations.}

\item{annealing}{boolean, apply annealing (alpha = 1/(i+1)) (default = TRUE).}
}
\value{
A new drtm with 'n_iter' times more iterations.
}
\description{
Minimize the gloabal energy of a drtm model.
}
\examples{
# Example model
m <- drt_import(
  system.file("example_i1000.RData", package = "drtplanr")
)

drt_iterate(m, 10)
}
