% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drtm.R
\name{drt_iterate}
\alias{drt_iterate}
\title{Iterate}
\usage{
drt_iterate(obj, n_iter)
}
\arguments{
\item{obj, }{drtm, a drtm model.}

\item{n_iter}{numeric, number of iterations.}
}
\value{
A new drtm with 'n_iter' times more iterations.
}
\description{
Minimize the gloabal energy of a drtm model.
}
\examples{
# Example model
m <- drt_import(
  system.file("Jegenstorf_i0.RData", package="drtplanr")
)

drt_iterate(m, 10)
}
