% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drtm.R
\name{drt_map}
\alias{drt_map}
\title{Plot station map}
\usage{
drt_map(obj)
}
\arguments{
\item{obj}{drtm, a drtm model.}
}
\value{
A mapview map object, containing the energy curve.
}
\description{
Plot station map
}
\examples{
# Example model
m <- drt_import(
  system.file("Jegenstorf_i1000.RData", package = "drtplanr")
)

drt_map(m)
}
