% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drtm.R
\name{drt_plot}
\alias{drt_plot}
\title{Plot energy curve}
\usage{
drt_plot(obj)
}
\arguments{
\item{obj}{drtm, a drtm model.}
}
\value{
A ggplot2 plor object, containing the energy curve.
}
\description{
Plot energy curve
}
\examples{
# Example model
m <- drt_import(
  system.file("Jegenstorf_i1000.RData", package = "drtplanr")
)

# Plot
drt_plot(m)
}
