% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drtm.R
\name{drt_save_graphics}
\alias{drt_save_graphics}
\alias{drt_save_graphics.drtm}
\title{Save graphics of model}
\usage{
drt_save_graphics(x, ...)

\method{drt_save_graphics}{drtm}(obj, path)
}
\arguments{
\item{obj}{drtm, a drtm model.}

\item{path}{character, path to file.}
}
\value{

}
\description{
Save graphics of model
}
