% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drtplanr-package.R
\docType{package}
\name{drtplanr-package}
\alias{drtplanr-package}
\alias{_PACKAGE}
\title{Demand-responsive transport planner}
\description{
Tool for placing virtual stations in demand-responsive transport systems in villages by defining and minimizing a global energy (\code{drtplanr}, name is inspired by \code{stplanr} <https://github.com/ropensci/stplanr>). The station locations are randomly initialized in the street network and iteratively optimized based on the reachable population in combination with walking and driving times.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/munterfinger/drtplanr}
  \item Report bugs at \url{https://github.com/munterfinger/drtplanr/issues}
}

}
\author{
Merlin Unterfinger - \email{info@munterfinger.ch}
}
\keyword{internal}
